# Optimize suboptimal code