--- a/hyprlock.conf
+++ b/hyprlock.conf
@@ -108,22 +108,7 @@ background {
     valign              = top
 }
 
-# WEATHERCAST & LOCATION
-label {
-    monitor     =
-    text        = cmd[update:1000] echo "$(bash $scrPath/weatherinfo.sh)"
-    
-    shadow_passes       = 1
-    shadow_boost        = 0.5
-    
-    color               = $fg0
-    font_size           = 11
-    font_family         = $font-text
 
-    position            = 0, -20
-    halign              = center
-    valign              = top
-}
 
 # INPUT FIELD
 input-field {
@@ -536,4 +521,19 @@ label {
     valign              = bottom
 }
 
+# WEATHERCAST & LOCATION
+label {
+    monitor     =
+    text        = cmd[update:1000] echo "$(bash $scrPath/weatherinfo.sh)"
+    
+    shadow_passes       = 1
+    shadow_boost        = 0.5
+    
+    color               = $fg0
+    font_size           = 11
+    font_family         = $font-text
 
+    position            = 0, -20
+    halign              = center
+    valign              = top
+}