The scripts were slow so a optimization
1)

Before:
./network.sh  0.04s user 0.02s system 1% cpu 3.832 total
After:
./network.sh  0.01s user 0.01s system 105% cpu 0.020 total

nmcli is heavy 
Changed it to iwget that comes with in wireless_tools

2)
Moved the weather to last to as it is only thing that could be slow due to user's internet if it would be at last it would load at last which means it wont affect the speed of hyprlock 

Great theme tho I love it :)