@tr1xem 2025-01-19
Closing